Python_Backend/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
datasets==3.1.0
Deprecated==1.2.15
dill==0.3.8
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.13.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.2.0
fuzzywuzzy==0.18.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
graphlib_backport==1.1.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kubernetes==31.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
# nvidia-cublas-cu11==11.11.3.6
# nvidia-cuda-cupti-cu11==11.8.87
# nvidia-cuda-nvrtc-cu11==11.8.89
# nvidia-cuda-runtime-cu11==11.8.89
# nvidia-cudnn-cu11==9.1.0.70
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.3.0.86
# nvidia-cusolver-cu11==11.4.1.48
# nvidia-cusparse-cu11==11.7.5.86
# nvidia-nccl-cu11==2.20.5
# nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
onnxruntime==1.16.3
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pillow==10.2.0
posthog==3.7.3
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.11
PyMySQL==1.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.2.1
# torch==2.4.1+cu118
# torchaudio==2.4.1+cu118
# torchvision==0.19.1+cu118
tqdm==4.67.1
transformers==4.46.3
# triton==3.0.0
typer==0.13.1
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
# uvloop==0.21.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
